// src/app/core/models/fileMetadataDto.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */

export interface FileMetadataDto {
  id: string;
  fileName: string;
  filePath: string;
  contentType: string;
  size: number;
  uploadDate: string;
  tenantId: string;
}