// src/app/core/models/documentRequirement.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */

export interface DocumentRequirement {
  documentType: string;
  documentName: string;
  requirementReason: string;
  isRequired: boolean;
  isUploaded: boolean;
  fileMetadataId: string;
  verificationStatus: string;
}