// src/app/core/models/documentVerificationStatus.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */

export type DocumentVerificationStatus = typeof DocumentVerificationStatus[keyof typeof DocumentVerificationStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DocumentVerificationStatus = {
  Pending: 0,
  Approved: 1,
  Rejected: 2,
  RequiresResubmission: 3,
} as const;